requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.10
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
pydantic==2.11.4
pydantic_core==2.33.2
pypiwin32==223
pyttsx3==2.98
pywin32==310
regex==2024.11.6
requests==2.32.3
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
